Convert the type tA2DP_STATUS into an enum
Bug: 331817295
Test: m com.android.btservices
Flag: EXEMPT mechanical refactor

+++ system/stack/a2dp/a2dp_sbc.cc
@@ -719,7 +719,8 @@ std::string A2DP_CodecInfoStringSbc(const uint8_t* p_codec_info) {
 
   a2dp_status = A2DP_ParseInfoSbc(&sbc_cie, p_codec_info, true);
   if (a2dp_status != A2DP_SUCCESS) {
-    res << "A2DP_ParseInfoSbc fail: " << loghex(a2dp_status);
+    res << "A2DP_ParseInfoSbc fail: "
+        << loghex(static_cast<uint8_t>(a2dp_status));
     return res.str();
   }
